Solve lots of backend errors , publish course , My course page for instructor

+++ server/controllers/Section.js
@@ -1,6 +1,6 @@
 const Section = require("../models/Section");
 const Course = require("../models/Course");
-
+const SubSection = require("../models/SubSection");
 exports.createSection = async (req, res) => {
     try{
         //data fetch
@@ -18,12 +18,12 @@ exports.createSection = async (req, res) => {
         const newSection = await Section.create({sectionName});
 
         //update course with section ObjectID
-        const updatedCourseDetails = await Course.findByIdAndUpdate(
+        const updatedCourse = await Course.findByIdAndUpdate(
                                             courseId,
                                             {
                                                 $push:{
                                                     courseContent:newSection._id,
-                                                }
+                                                },
                                             },
                                             {new:true},
                                         ).populate({
@@ -40,7 +40,7 @@ exports.createSection = async (req, res) => {
         return res.status(200).json({
             success:true,
             message:'Section created successfully',
-            updatedCourseDetails,
+            updatedCourse,
         })
     }
     catch(error) {
@@ -56,7 +56,7 @@ exports.updateSection = async (req,res) => {
     try {
 
         //data input
-        const {sectionName, sectionId} = req.body;
+        const {sectionName, sectionId, courseId} = req.body;
 
         //data validation
         if(!sectionName || !sectionId) {
@@ -69,10 +69,20 @@ exports.updateSection = async (req,res) => {
         //update data
         const section = await Section.findByIdAndUpdate(sectionId, {sectionName}, {new:true});
 
+        const course = await Course.findById(courseId)
+		.populate({
+			path:"courseContent",
+			populate:{
+				path:"subSection",
+			},
+		})
+		.exec();
+
         //return res
         return res.status(200).json({
             success:true,
-            message:'Section Updated Successfully',
+            message:section,
+            data:course,
         });
 
     }
@@ -86,28 +96,49 @@ exports.updateSection = async (req,res) => {
 };
 
 
-exports.deleteSection = async (req,res) => {
-    try {
-        //get ID - assuming that we are sending ID in params
-        const {sectionId} = req.params
+// DELETE a section
+exports.deleteSection = async (req, res) => {
+	try {
 
-        //use findByIdandDelete
-        await Section.findByIdAndDelete(sectionId);
+		const { sectionId, courseId }  = req.body;
+		await Course.findByIdAndUpdate(courseId, {
+			$pull: {
+				courseContent: sectionId,
+			}
+		})
+		const section = await Section.findById(sectionId);
+		console.log(sectionId, courseId);
+		if(!section) {
+			return res.status(404).json({
+				success:false,
+				message:"Section not Found",
+			})
+		}
 
-        //TODO[Testing]: do we need to delete the entry from the course schema ??
-        
-        //return response
-        return res.status(200).json({
-            success:true,
-            message:"Section Deleted Successfully",
-        })
+		//delete sub section
+		await SubSection.deleteMany({_id: {$in: section.subSection}});
 
-    }
-    catch(error) {
-        return res.status(500).json({
-            success:false,
-            message:"Unable to delete Section, please try again",
-            error:error.message,
-        });
-    }
-}
+		await Section.findByIdAndDelete(sectionId);
+
+		//find the updated course and return 
+		const course = await Course.findById(courseId).populate({
+			path:"courseContent",
+			populate: {
+				path: "subSection"
+			}
+		})
+		.exec();
+
+		res.status(200).json({
+			success:true,
+			message:"Section deleted",
+			data:course
+		});
+	} catch (error) {
+		console.error("Error deleting section:", error);
+		res.status(500).json({
+			success: false,
+			message: "Internal server error",
+		});
+	}
+};
